requirements.txt
PySide6>=6.6.0
google-genai>=1.0.0
requests>=2.31.0
pyinstaller>=6.0.0
